requirements.txt
certifi==2020.12.5
llvmlite==0.36.0
mkl-fft==1.3.0
mkl-random==1.2.1
mkl-service==2.3.0
numba==0.53.1
numpy==1.20.1
Pillow==8.2.0
pip==21.0.1
scipy==1.6.2
setuptools==52.0.0.post20210125
six==1.15.0
wheel==0.36.2
